requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
deepface>=0.0.79
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
gdown>=4.7.0